feat: Add getBattlesByVillainName endpoint
The battlesController and battlesRoutes files have been modified to add a new endpoint for retrieving battles by villain name. This allows users to get battles where a specific villain is either the winner or the loser. This change enhances the functionality of the codebase by providing a new query option for battles.